+++ src/index.ts
@@ -2,9 +2,10 @@
  *
  * index.ts: homebridge-august.
  */
-import { PLATFORM_NAME, PLUGIN_NAME } from './settings.js';
-import { API } from 'homebridge';
 import { AugustPlatform } from './platform.js';
+import { PLATFORM_NAME, PLUGIN_NAME } from './settings.js';
+
+import type { API } from 'homebridge';
 
 // Register our platform with homebridge.
 export default (api: API): void => {